dashboard/requirements.txt
flask==3.1.0
docker==6.1.3
pyyaml==6.0.1
python-dotenv==1.0.0